MenuPlan 추가 기능 변경 - 기존 기능에 추가로 pos를 입력할 수 있도록 변경

+++ cafeteria-manage-api/src/main/java/com/poppo/dallab/cafeteria/domain/MenuPlan.java
@@ -25,5 +25,5 @@ public class MenuPlan {
 
     private Long menuId;
 
-    private Integer pos;
+    private Double pos;
 }
